+++ src/Doctrine/ORM/Locking/LockException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace FactoryGirl\Provider\Doctrine\ORM\Locking;
-
-use FactoryGirl\Provider\Doctrine\Exception;
-
-class LockException extends \Exception implements Exception
-{
-}